moves rcmg_draw_fn logic into JointModel
