+++ setup.py
@@ -30,7 +30,7 @@ def find_data_files(package_dir, patterns, excludes=()):
 setuptools.setup(
     name="x_xy",
     packages=setuptools.find_packages(),
-    version="0.2.10",
+    version="0.2.11",
     package_data={"x_xy": find_data_files("x_xy", patterns=["*.xml"])},
     include_package_data=True,
     install_requires=[
