+++ x_xy/algorithms/__init__.py
@@ -1,13 +1,14 @@
+from . import dynamics, jcalc, kinematics, sensors
 from .dynamics import compute_mass_matrix, forward_dynamics, inverse_dynamics, step
-from .jcalc import jcalc_motion, jcalc_tau, jcalc_transform
+from .jcalc import (
+    JointModel,
+    RCMG_Config,
+    jcalc_motion,
+    jcalc_tau,
+    jcalc_transform,
+    register_new_joint_type,
+)
 from .kinematics import forward_kinematics, forward_kinematics_transforms
 from .random import random_angle_over_time, random_position_over_time
-from .rcmg import (
-    FINALIZE_FN,
-    SETUP_FN,
-    Generator,
-    RCMG_Config,
-    batch_generator,
-    build_generator,
-)
+from .rcmg import FINALIZE_FN, SETUP_FN, Generator, batch_generator, build_generator
 from .sensors import accelerometer, add_noise_bias, gyroscope, imu, rel_pose
